newton: Fix bug when cfg.simplify=true
The functor implemented in e.g. tmb_examples/TMbad/spa_gauss.cpp works specifically for ad_aug types, but is also used for double types by the newton operator. A subtle side effect is that the evaluation may add to the global tape even if such tape doesn't exist. This problem happens when the functor uses operators that (by mistake) do not test for constant input.