Translation unit changes:
Packages can use 'TMB.cpp' and 'TMB.h' where the former 'precompiles'
a substantial portion of TMB. This commit gives better control with
how much is precompiled.

New flags 'TMB_PRECOMPILE_ATOMICS'/'HAVE_PRECOMPILED_ATOMICS' can be
used to include all atomic functions in precompilation, and to enable
the corresponding headers.

When *not* setting these flags, the precompiled unit becomes much
smaller and faster to compile. Atomic functions can still be used
(new) but will be compiled on demand.

The new flag 'TMB_PRECOMPILE_ATOMICS' replaces the old flag
'TMB_PRECOMPILE' (so that downstream use of 'TMB_PRECOMPILE' is
ignored and the default behaviour is now to exclude atomics from
precompilation when used by a package).

By contrast, the function 'TMB::precompile()' uses the new flags in
order to precompile as much as possible (like it used to do).