TMBad src update:
checkpoint fix higher order retaping.

Retaping case was originally intended for unique instances where the
'dtab' was not supposed to be shared. It could then be assumed that
the forward pass would take care of rebuilding 'dtab' when
necessary. We don't want that restriction anymore.

In addition, add unique identifier for the tape optimizer.